Securing /api endpoint with `secure_api` config
The /api namespace would benefit from security control. The actual configuring
of ports and certificates is done in Torquebox configuration, but we can ensure
that communications are secure in each call to /api endpoints. This change adds
a config property called `secure_api` that, when enabled, returns a 401
(Unauthorized) when an /api endpoint is called without HTTPS. By default,
this is disabled (same as current behavior).

Fixes https://tickets.puppetlabs.com/browse/RAZOR-487